feat/studyViewerExportToWorkspace: working
