+++ src/StudyViewer/SingleStudyViewer.jsx
@@ -8,7 +8,7 @@ import { FileOutlined, FilePdfOutlined, LinkOutlined } from '@ant-design/icons';
 import BackLink from '../components/BackLink';
 import { humanFileSize } from '../utils.js';
 import {
-  ReduxStudyDetails, fetchDataset, fetchFiles, resetMultipleStudyData, fetchStudyViewerConfig,
+  ReduxStudyDetails, fetchDataset, fetchFiles, resetMultipleStudyData, fetchStudyViewerConfig, ReduxExportToWorkspace,
 } from './reduxer';
 import getReduxStore from '../reduxStore';
 import './StudyViewer.css';
@@ -43,8 +43,29 @@ class SingleStudyViewer extends React.Component {
     this.state = {
       dataType: undefined,
       rowAccessor: undefined,
+      exportToWorkspace: {},
+      exportingPFBToWorkspace: false,
     };
   }
+
+  exportToWorkspace = (buttonConfig) => {
+    this.setState({
+      exportToWorkspace: {...buttonConfig},
+    });
+  };
+
+  exportingPFBToWorkspaceStateChange = (stateChange) => {
+    let tempStateChange = {
+        exportingPFBToWorkspace: stateChange,
+      };
+
+    // if set to false clear exportToWorkspace
+    if (!stateChange) {
+      tempStateChange.exportToWorkspace = {};
+    }
+
+    this.setState(tempStateChange);
+  };
 
   static getDerivedStateFromProps(nextProps, prevState) {
     const newState = {};
@@ -111,11 +132,13 @@ class SingleStudyViewer extends React.Component {
           </div>
           <div className='study-viewer__details'>
             <ReduxStudyDetails
-              data={dataset}
-              fileData={this.props.fileData}
-              studyViewerConfig={studyViewerConfig}
-              isSingleItemView={false}
-            />
+                data={dataset}
+                fileData={this.props.fileData}
+                studyViewerConfig={studyViewerConfig}
+                isSingleItemView={false}
+                exportToWorkspaceAction={this.exportToWorkspace}
+                exportToWorkspaceEnabled={!this.state.exportingPFBToWorkspace}
+              />
             <div className='study-viewer__details-sidebar'>
               <Space direction='vertical' style={{ width: '100%' }}>
                 {(sideBoxesConfig && sideBoxesConfig.length > 0)
@@ -158,6 +181,11 @@ class SingleStudyViewer extends React.Component {
             </div>
           </div>
         </Space>
+        <ReduxExportToWorkspace
+            exportToWorkspaceAction={this.state.exportToWorkspace}
+            exportingPFBToWorkspaceStateChange={this.exportingPFBToWorkspaceStateChange}
+            exportingPFBToWorkspace={this.state.exportingPFBToWorkspace}
+          ></ReduxExportToWorkspace>
       </div>
     );
   }
