+++ src/StudyViewer/reduxer.js
@@ -3,6 +3,7 @@ import _ from 'lodash';
 import { withRouter } from 'react-router-dom';
 import StudyDetails from './StudyDetails';
 import StudyViewer from './StudyViewer';
+import ExportToWorkspace from './ExportToWorkspace';
 import SingleStudyViewer from './SingleStudyViewer';
 import { guppyGraphQLUrl, studyViewerConfig, requestorPath } from '../localconf';
 import { fetchWithCreds } from '../actions';
@@ -246,16 +247,9 @@ export const ReduxStudyViewer = (() => {
     docData: state.study.docData,
     fileData: state.study.fileData,
     noConfigError: state.study.noConfigError,
-    job: state.analysis.job,
-  });
-  const mapDispatchToProps = (dispatch) => ({
-    submitJob: (body) => dispatch(dispatchJob(body)),
-    checkJobStatus: () => dispatch(checkJob()),
-    fetchJobResult: (jobId) => dispatch(fetchJobResult(jobId)),
-    resetJobState: () => dispatch(resetJobState()),
   });
 
-  return withRouter(connect(mapStateToProps, mapDispatchToProps)(StudyViewer));
+  return withRouter(connect(mapStateToProps)(StudyViewer));
 })();
 
 export const ReduxSingleStudyViewer = (() => {
@@ -264,6 +258,13 @@ export const ReduxSingleStudyViewer = (() => {
     docData: state.study.docData,
     fileData: state.study.fileData,
     noConfigError: state.study.noConfigError,
+  });
+
+  return withRouter(connect(mapStateToProps)(SingleStudyViewer));
+})();
+
+export const ReduxExportToWorkspace = (() => {
+  const mapStateToProps = (state) => ({
     job: state.analysis.job,
   });
   const mapDispatchToProps = (dispatch) => ({
@@ -273,5 +274,5 @@ export const ReduxSingleStudyViewer = (() => {
     resetJobState: () => dispatch(resetJobState()),
   });
 
-  return withRouter(connect(mapStateToProps, mapDispatchToProps)(SingleStudyViewer));
+  return withRouter(connect(mapStateToProps, mapDispatchToProps)(ExportToWorkspace));
 })();
